Fix for Fabric com.tastory.TastoryApp.iOS issue #1 - Markup[0] nil
If for some reason Moment[0] was deemed ready, StoryOperation.allMedia/nextmedia will skip retrieving Moment[0]. Problem is that we currently rely on Moment[0] retrieval to signify Batch PFQuery for Moment and Markup Parse data structures. Skipping Moment[0] causes the Parse Moments and Markup for some later Moment to be skipped/missing. Once an access is attempted on them, we will log and crash.

Fix is to move the BatchRetrieve/Query out from Moment[0] as it's own instance of operation. So the first time a StoryOperation.allMedia/nextMedia is executed, it'll first see if a batch is needed. If needed the operation will just end there. Only until the next cycle in allMedia, or the following nextMedia request will Moment[0] media be attempted to be retrieved.