Fix problem where in-app camera video doesn't loop in MarkupVC
Problem is because MarkupVC's AVPlayer already started by the time CameraVC's viewDidDisappear gets called. In there it in turns calls SwiftyCamVC.viewDidDisappear which stops the AVCaptureSession. This somehow also stops the MarkupVC's avPlayer from playing.

Workaround is to first manually stop the Capture Session manually before presenting the MarkupVC, or all subsequent VCs from the CameraVC for that matter.